+++ test/unit_tests/braket/default_simulator/test_state_vector_simulation.py
@@ -17,7 +17,7 @@ import numpy as np
 import pytest
 
 from braket.default_simulator import gate_operations, observables
-from braket.default_simulator.statevector_simulation import StateVectorSimulation
+from braket.default_simulator.state_vector_simulation import StateVectorSimulation
 
 evolve_testdata = [
     ([gate_operations.Hadamard([0])], 1, [0.70710678, 0.70710678], [0.5, 0.5]),
